Fix login: PBKDF2, CLI (seed-admin/set-password), user_loader, CI SQLite
